Add example for progress tracking
