+++ Polygon_mesh_processing/doc/Polygon_mesh_processing/examples.txt
@@ -23,6 +23,7 @@
 \example Poisson_surface_reconstruction_3/poisson_reconstruction_example.cpp
 \example Polygon_mesh_processing/corefinement_difference_remeshed.cpp
 \example Polygon_mesh_processing/corefinement_mesh_union.cpp
+\example Polygon_mesh_processing/corefinement_mesh_union_progress.cpp
 \example Polygon_mesh_processing/corefinement_OM_union.cpp
 \example Polygon_mesh_processing/corefinement_mesh_union_and_intersection.cpp
 \example Polygon_mesh_processing/corefinement_consecutive_bool_op.cpp
