+++ Polygon_mesh_processing/examples/Polygon_mesh_processing/CMakeLists.txt
@@ -78,6 +78,7 @@ create_single_source_cgal_program("corefinement_SM.cpp")
 create_single_source_cgal_program("corefinement_consecutive_bool_op.cpp")
 create_single_source_cgal_program("corefinement_difference_remeshed.cpp")
 create_single_source_cgal_program("corefinement_mesh_union.cpp")
+create_single_source_cgal_program("corefinement_mesh_union_progress.cpp")
 create_single_source_cgal_program(
   "corefinement_mesh_union_and_intersection.cpp")
 create_single_source_cgal_program("corefinement_mesh_union_with_attributes.cpp")
